+++ lib/policy/checkresult.go
@@ -1,4 +1,4 @@
-package lib
+package policy
 
 import (
 	"log/slog"
@@ -15,11 +15,4 @@ func (cr CheckResult) LogValue() slog.Value {
 	return slog.GroupValue(
 		slog.String("name", cr.Name),
 		slog.String("rule", string(cr.Rule)))
-}
-
-func cr(name string, rule config.Rule) CheckResult {
-	return CheckResult{
-		Name: name,
-		Rule: rule,
-	}
 }